# [FIX] product_margin: group_by products margins computations