Group by products margins computations to reduce SQL queries sent
to postgre server, improving overall performances.


##### Speedup

Client DB: 4k active products, 2.5k active templates, 140k account_moves, 564k account_move_lines
Accounting > Reporting > Product margins.

| Nb active products | group_by product category current | group_by product category after PR |
| :------------------: | :-----------------------------------: | :-----------------------------------: |
| 500                       | 48s                                                  | 1.30s                                               |
| 1000                     | 1min30s                                          |  3.30s                                              |
| 2000                     | 3min20s                                          |  6.20s                                              |
| 4117                     | 6min36s                                          | 10.30s                                           |




--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr