@Aurelienvd 2021-06-01
> 
> 
> all done?

Yes. I looped over `self.env.cr.fetchall()`. Since I had no guarantee that the first loop will return the same product_ids than the second one I had to "precompute" additional fields in the first loop. I also looped once on `self.ids` for the write but overall there are less iterations.