@Aurelienvd 2021-06-04
@qdp-odoo Do you think it is OK now ?