@Aurelienvd 2021-07-09
@qdp-odoo r+ ? (I don't have r+ rights :sweat_smile: )