@Aurelienvd 2021-07-22
@odoo/rd-accounting Could this PR be r+ ? approved by qdp but I don't have r+ rights 😃